Removed sandbox and implemented new MediaItem and Video.js
